requirements.txt
langchain
requests
python-dotenv
gradio
tavily-python
langchain-community